Move parser types into parser_types.hpp.
